fix: widen PostHog personal key detection
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
