+++ src/agentsweep/scanner.py
@@ -269,7 +269,7 @@ RULES: list[tuple[str, str, re.Pattern]] = [
     ('planetscale-password', 'PlanetScale database password',
         re.compile('\\bpscale_pw_[\\w=.\\-]{32,64}(?![\\w=.\\-])')),
     ('posthog-personal-api-key', 'PostHog personal API key',
-        re.compile('\\bphx_[2-9A-HJ-NP-Za-km-z]{48,49}\\b')),
+        re.compile('\\bphx_[A-Za-z0-9]{43,49}\\b')),
     ('postman-api-key', 'Postman API key',
         re.compile('\\bPMAK-[a-fA-F0-9]{24}-[a-fA-F0-9]{34}\\b')),
     ('prefect-api-key', 'Prefect API key',
@@ -837,7 +837,7 @@ ROTATION_GUIDANCE: dict[str, str] = {
     'planetscale-api-token': 'Revoke: https://app.planetscale.com (Organization settings > Service tokens)',
     'planetscale-oauth-token': 'Revoke: https://app.planetscale.com (Organization settings > OAuth applications)',
     'planetscale-password': 'Rotate: pscale password delete / pscale password create, or https://app.planetscale.com (database > Settings > Passwords)',
-    'posthog-personal-api-key': 'Revoke: PostHog Settings > Personal API keys (https://us.posthog.com/settings/user-api-keys)',
+    'posthog-personal-api-key': 'Revoke: PostHog Settings > Personal API keys in your PostHog instance (US/EU cloud or self-hosted).',
     'postman-api-key': 'Revoke: https://go.postman.co/settings/me/api-keys',
     'prefect-api-key': 'Revoke: https://app.prefect.cloud/my/api-keys',
     'pulumi-access-token': 'Revoke: https://app.pulumi.com/account/tokens',
